cleanup better clioin make support

+++ cpp/lang_compare_server.cc
@@ -18,7 +18,7 @@ using namespace std;
 string applyXorCipher(string const &key, string const &cs) {
     string res = cs;
     auto mod = key.size() / sizeof(char);
-    for (int i = 0; i < cs.size(); i++) {
+    for (u_long i = 0; i < cs.size(); i++) {
         auto c = key[i % mod];
         res[i] = cs[i] ^ c;
     }
@@ -29,7 +29,6 @@ string applyXorCipher(string const &key, string const &cs) {
 // Logic and data behind the server's behavior.
 class LangCompareServiceImpl final : public LangCompare::Service {
     Status XorCipher(ServerContext* context, const XorCipherRequest* request, XorCipherReply* reply) override {
-//        cout << "key: " << request->key() << ", in_str: " << request->in_str() << "\n";
         auto res = applyXorCipher(request->key(), request->in_str());
         reply->set_out_str(res);
         return Status::OK;
@@ -41,11 +40,13 @@ void RunServer() {
   LangCompareServiceImpl service;
 
   ServerBuilder builder;
+
   // Listen on the given address without any authentication mechanism.
   builder.AddListeningPort(server_address, grpc::InsecureServerCredentials());
-  // Register "service" as the instance through which we'll communicate with
-  // clients. In this case it corresponds to an *synchronous* service.
+
+  // Register "service" as the instance through which we'll communicate with clients.
   builder.RegisterService(&service);
+
   // Finally assemble the server.
   unique_ptr<Server> server(builder.BuildAndStart());
   cout << "C++ lang-compare Server listening on " << server_address << endl;
